+++ src/store/change_once_subscribe/index.d.ts
@@ -1,11 +1,12 @@
 import type { Readable, Subscriber, Unsubscriber } from 'svelte/store'
+import type { StoresValues } from '../_types'
 /**
  * Calls the given fn the next time the value of the store changes, then unsubscribes.
  */
 export declare function change_once_subscribe<
-	Val extends unknown = unknown
+	Store extends Readable<unknown>
 >(
-	store:Readable<Val>,
-	fn:Subscriber<Val>
+	store:Store,
+	fn:Subscriber<StoresValues<Store>>
 ):Unsubscriber
 export { change_once_subscribe as subscribe__change__once }