+++ src/store/clear_store/index.d.ts
@@ -3,9 +3,9 @@ import type { Writable } from 'svelte/store'
  * Sets each store with val
  */
 export declare function clear_store<
-	Val extends unknown = unknown
+	Store extends Writable<unknown>
 >(
-	stores:Writable<Val>|Writable<Val>[],
-	val:Val
+	stores:Store|Store[],
+	val:Writable<Store>
 ):void
 export { clear_store as clear__store }