+++ src/store/clear_store_/index.d.ts
@@ -1,12 +1,13 @@
 import type { Writable } from 'svelte/store'
+import type { StoresValues } from '../_types'
 /**
  * Returns a function that [clear_store](#clear_store).
  */
 export declare function clear_store_<
-	Val extends unknown = unknown
+	Store extends Writable<unknown>
 >(
-	stores:Writable<Val>|Writable<Val>[],
-	value:Val
+	stores:Store|Store[],
+	value:StoresValues<Store>
 ):clear_store_T
 export declare type clear_store_T = ()=>void
 export {