+++ src/store/mix_set_readable_/index.d.ts
@@ -1,12 +1,13 @@
+import type { StoresValues } from '../_types'
 import type { mix_set_readable_I } from '../mix_set_readable'
 import type { Readable } from '../readable'
 import type { Writable_set_T } from '../writable'
 export declare function mix_set_readable_<
-	Val extends unknown = unknown
+	Store extends Readable<unknown>
 >(
-	store:Readable<Val>,
-	set?:Writable_set_T<Val>
-):mix_set_readable_T<Val>
+	store:Store,
+	set?:Writable_set_T<StoresValues<Store>>
+):mix_set_readable_T<StoresValues<Store>>
 export interface mix_set_readable_T<
 	Val extends unknown = unknown
 > extends mix_set_readable_I<Val> {