+++ src/store/multi_subscribe/index.d.ts
@@ -1,12 +1,12 @@
 import type { Readable, Unsubscriber } from 'svelte/store'
+import type { StoresValues } from '../_types'
 /**
  * Subscribes to multiple stores. The subscriber fn is called when any of the store_a changes.
  */
 export declare function multi_subscribe<
-	Val extends unknown,
-	Store extends Readable<any> = Readable<any>
+	Store extends Readable<unknown>
 >(
 	store_a:Store[],
-	fn:((val_a:Val[])=>void)
+	fn:((val_a:StoresValues<Store>[])=>void)
 ):Unsubscriber
 export { multi_subscribe as subscribe__multi }