+++ src/store/set_store_/index.d.ts
@@ -1,14 +1,15 @@
 import type { Writable } from 'svelte/store'
+import type { StoresValues } from '../_types'
 /**
  * Returns a function to set the given store using the value returned by `setter`.
  * This is useful in conjunction with [subscribe](#subscribe).
  */
 export declare function set_store_<
-	Val extends unknown = unknown
+	Store extends Writable<unknown>
 >(
-	store:Writable<Val>,
-	setter?:(v:Val)=>Val
-):set_store_T<Val>
+	store:Store,
+	setter?:(v:StoresValues<Store>)=>StoresValues<Store>
+):set_store_T<StoresValues<Store>>
 export declare type set_store_T<
 	Val extends unknown = unknown
 > = (val:Val)=>void