+++ src/store/subscribe/index.d.ts
@@ -1,12 +1,13 @@
 import type { Readable, Subscriber, Unsubscriber } from 'svelte/store'
+import type { StoresValues } from '../_types'
 import type { Invalidator } from '../_types'
 /**
  * Delegates to store.subscribe
  */
 export declare function subscribe<
-	Val extends unknown = unknown
+	Store extends Readable<unknown>
 >(
-	store:Readable<Val>,
-	run:Subscriber<Val>,
-	invalidate?:Invalidator<Val>
+	store:Store,
+	run:Subscriber<StoresValues<Store>>,
+	invalidate?:Invalidator<StoresValues<Store>>
 ):Unsubscriber